ai-service/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
pillow==10.4.0
imagehash==4.3.1
numpy==1.26.4
scikit-learn==1.5.2
torch==2.4.1
torchvision==0.19.1
timm==1.0.9
pymongo==4.8.0
python-multipart==0.0.12
loguru==0.7.2
python-dotenv==1.0.1
